+++ src/store/actions/index.js
@@ -1,5 +1,4 @@
 /* eslint-disable camelcase */
-import set_sidebar_items from './set_sidebar_items';
 import set_active_item from './set_active_item';
 import init_state from './init_state';
 import set_active_key_result from './set_active_key_result';
@@ -11,7 +10,6 @@ import set_user from './set_user';
 import update_preferences from './update_preferences';
 
 export default {
-  set_sidebar_items,
   set_active_item,
   set_active_key_result,
   set_active_objective,